Removed deprecated classes. Removed further unused classes and features.
