--- framework/application/core/src/main/java/org/ldp4j/application/template/InvalidAttachmentDefinition.java
@@ -1,44 +1,0 @@
-/**
- * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
- *   This file is part of the LDP4j Project:
- *     http://www.ldp4j.org/
- *
- *   Center for Open Middleware
- *     http://www.centeropenmiddleware.com/
- * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
- *   Copyright (C) 2014 Center for Open Middleware.
- * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *             http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
- *   Artifact    : org.ldp4j.framework:ldp4j-application-core:1.0.0-SNAPSHOT
- *   Bundle      : ldp4j-application-core-1.0.0-SNAPSHOT.jar
- * #-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=#
- */
-package org.ldp4j.application.template;
-
-class InvalidAttachmentDefinition extends TemplateLoadingRuntimeException {
-
-	private static final long serialVersionUID = 5228332100267319288L;
-
-	private final String attachmentId;
-
-	public InvalidAttachmentDefinition(String templateId, String attachmentId, String message, Throwable t) {
-		super(templateId,message,t);
-		this.attachmentId = attachmentId;
-	}
-
-	public String attachmentId() {
-		return attachmentId;
-	}
-
-}
